refactoring, create TwitSerfice interface
